RK4 Integration Added
I have added the option of RK4 for numerical integration and I have also added a checkbox in the initial form to select whether RK4 is used or not. Minor adjustments have been made regarding the range of the time step in the form because with a time step greater than 300ms the numerical integration gets too faulty as the mechanical energy of the pendulums steadily increases while it should remain somewhat constant.